Remove opensanctions.model, make CLI start again after that
