--- opensanctions/model/__init__.py
@@ -1,7 +1,0 @@
-from opensanctions.model.base import db
-from opensanctions.model.statement import Statement
-from opensanctions.model.resource import Resource
-from opensanctions.model.issue import Issue
-
-
-__all__ = ["db", "Statement", "Issue"]
